Move multizone e2e to sig scheduling path
