+++ hack/.golint_failures
@@ -788,7 +788,6 @@ test/e2e/instrumentation/monitoring
 test/e2e/kubectl
 test/e2e/lifecycle
 test/e2e/lifecycle/bootstrap
-test/e2e/multicluster
 test/e2e/network
 test/e2e/node
 test/e2e/scalability
