+++ test/e2e/scheduling/ubernetes_lite.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package multicluster
+package scheduling
 
 import (
 	"fmt"
